util: Move android utils and native_app_glue
Move GetIntentExtra, DestroyActivity and the native_app_glue library to
gfxrecon_util. This simplifies dependencies and exposes Android util
functions to the test targets.
