+++ test/test_apps/external-memory-fd/import/CMakeLists.txt
@@ -40,7 +40,6 @@ target_link_libraries(gfxrecon-testapp-external-memory-fd-import
         gfxrecon_decode
         gfxrecon_graphics
         gfxrecon_format
-        gfxrecon_util
         platform_specific
         gfxrecon-testapp-base)
