+++ tools/optimize/CMakeLists.txt
@@ -62,7 +62,6 @@ target_link_libraries(gfxrecon-optimize
                           gfxrecon_decode
                           gfxrecon_graphics
                           gfxrecon_format
-                          gfxrecon_util
                           platform_specific
                           $<$<BOOL:${D3D12_SUPPORT}>:d3d12.lib>
                           $<$<BOOL:${D3D12_SUPPORT}>:dxgi.lib>
