+++ tools/replay/CMakeLists.txt
@@ -46,7 +46,6 @@ target_link_libraries(gfxrecon-replay
                           gfxrecon_decode
                           gfxrecon_graphics
                           gfxrecon_format
-                          gfxrecon_util
                           platform_specific
                           $<$<BOOL:${D3D12_SUPPORT}>:d3d12.lib>
                           $<$<BOOL:${D3D12_SUPPORT}>:dxgi.lib>
